made file organization cleaner
