+++ client/src/components/Account/Account.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Account() {
+  return (
+    <div>
+      <p>My Account</p>
+    </div>
+  );
+}
