+++ client/src/components/Account/List/List.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function List() {
+  return (
+    <div>
+      <p>My List</p>
+    </div>
+  );
+}
