+++ client/src/components/Account/Profile/Profile.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default function Profile() {
+  return (
+    <div>
+      <p>My Profile</p>
+    </div>
+  );
+}
