+++ client/src/components/Firebase/Exports.js
@@ -1,0 +1,4 @@
+import FirebaseContext, { withFirebase } from "./Context";
+import Firebase from "./Firebase";
+
+export { Firebase, FirebaseContext, withFirebase };
