+++ client/src/components/Home/Header/Header.js
@@ -1,0 +1,14 @@
+import React, { Component } from 'react'
+import './Header.scss'
+
+class Header extends Component {
+  render() {
+    return (
+      <div className="header">
+        <h1>Hot Code</h1>
+      </div>
+    )
+  }
+}
+
+export default Header;
