+++ client/src/components/Home/Header/Header.scss
@@ -1,0 +1,8 @@
+.header{
+  display:flex;
+  height:100px;
+  width:100%;
+  padding:10px;
+  border: 2px solid black;
+  justify-content: center;
+}
