+++ client/src/components/PasswordForget/PasswordForget.js
@@ -1,0 +1,11 @@
+import React, { Component } from "react";
+
+export default class PasswordForget extends Component {
+  render() {
+    return (
+      <div>
+        <p>passwordForget</p>
+      </div>
+    );
+  }
+}
