+++ client/src/components/SignOut/SignOut.js
@@ -1,0 +1,10 @@
+import React from "react";
+import { withFirebase } from "../Firebase/Exports";
+
+const SignOutButton = ({ firebase }) => (
+  <button type="button" onClick={firebase.doSignOut}>
+    Sign Out
+  </button>
+);
+
+export default withFirebase(SignOutButton);
