--- client/src/components/cards/card.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "./card.scss";
-
-function Card() {
-  return (
-    <div className="card">
-      <p>Props.Image</p>
-      <p>Props.Stars</p>
-      <p>Props.user</p>
-    </div>
-  );
-}
-
-export default Card;
