--- client/src/components/firebase/index.js
@@ -1,4 +1,0 @@
-import FirebaseContext, { withFirebase } from "./context";
-import Firebase from "./firebase";
-
-export { Firebase, FirebaseContext, withFirebase };
