--- client/src/components/nav/nav.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-import './nav.scss'
-
-class Nav extends Component {
-  render() {
-    return (
-      <div className="navigation">
-        <h1>Hot Code</h1>
-
-      </div>
-    )
-  }
-}
-
-export default Nav;
