--- client/src/components/nav/nav.scss
@@ -1,8 +1,0 @@
-.navigation{
-  display:flex;
-  height:100px;
-  width:100%;
-  padding:10px;
-  border: 2px solid black;
-  justify-content: center;
-}
