--- client/src/components/passwordForget/passwordForget.js
@@ -1,11 +1,0 @@
-import React, { Component } from "react";
-
-export default class passwordForget extends Component {
-  render() {
-    return (
-      <div>
-        <p>passwordForget</p>
-      </div>
-    );
-  }
-}
