--- client/src/components/signOut/signOut.js
@@ -1,10 +1,0 @@
-import React from "react";
-import { withFirebase } from "../../components/firebase/index";
-
-const SignOutButton = ({ firebase }) => (
-  <button type="button" onClick={firebase.doSignOut}>
-    Sign Out
-  </button>
-);
-
-export default withFirebase(SignOutButton);
