+++ client/src/index.js
@@ -1,10 +1,10 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./styles/_globals.scss";
-import App from "./components/app/App";
+import App from "./App";
 import ApolloClient from "apollo-boost";
 import { ApolloProvider } from "react-apollo";
-import { FirebaseContext, Firebase } from "./components/firebase/index.js";
+import { FirebaseContext, Firebase } from "./components/Firebase/Exports";
 
 const client = new ApolloClient({
   uri: "https://stark-meadow-13883.herokuapp.com/"
