+++ client/src/reactRouter/reactRouter.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import * as ROUTES from "../constants/routes";
-import SignOutButton from "../components/signOut/signOut";
+import SignOutButton from "../components/SignOut/SignOut";
 
 const Navigation = ({ authUser }) => (
   <div>{authUser ? <NavigationAuth /> : <NavigationNonAuth />}</div>
@@ -14,16 +14,10 @@ const NavigationAuth = () => {
         <Link to={ROUTES.SIGN_IN}>Sign In</Link>
       </li>
       <li>
-        <Link to={ROUTES.LANDING}>Landing Page</Link>
-      </li>
-      <li>
-        <Link to={ROUTES.HOME}>Home Page</Link>
+        <Link to={ROUTES.HOME}>Home</Link>
       </li>
       <li>
         <Link to={ROUTES.ACCOUNT}>My Account</Link>
-      </li>
-      <li>
-        <Link to={ROUTES.ADMIN}>Admin</Link>
       </li>
       <SignOutButton />
     </ul>
@@ -33,7 +27,7 @@ const NavigationNonAuth = () => {
   return (
     <ul>
       <li>
-        <Link to={ROUTES.LANDING}>Landing</Link>
+        <Link to={ROUTES.HOME}>Home</Link>
       </li>
       <Link to={ROUTES.SIGN_IN}>Sign In</Link>
     </ul>
