+++ client/src/styles/card.scss
