--- client/src/views/adminPage/adminPage.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function adminPage() {
-  return (
-    <div>
-      <p>Admin Page</p>
-    </div>
-  )
-}
