--- client/src/views/homePage/homePage.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function homePage() {
-	return (
-		<div>
-			<h1>Hi</h1>
-		</div>
-	);
-}
