--- client/src/views/loginPage/login.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class Login extends Component {
-	render() {
-		return <div />;
-	}
-}
-
-export default Login;
